move specialization into the header (avoid superflous include)
