+++ Number_types/include/CGAL/Lazy_exact_nt.h
@@ -34,6 +34,8 @@
 #include <CGAL/Kernel/mpl.h>
 #include <CGAL/tss.h>
 #include <CGAL/type_traits.h>
+#include <CGAL/Number_types/internal/Exact_type_selector.h>
+
 
 #include<type_traits>
 
@@ -1427,7 +1429,24 @@ class Modular_traits<Lazy_exact_nt<ET> >
 #undef CGAL_int
 #undef CGAL_To_interval
 
-} //namespace CGAL
+namespace internal {
+
+template < typename ET >
+struct Exact_field_selector<Lazy_exact_nt<ET> >
+: Exact_field_selector<ET>
+{
+  // We have a choice here :
+  // - using ET gets rid of the DAG computation as well as redoing the interval
+  // - using Lazy_exact_nt<ET> might use sharper intervals.
+  // typedef ET  Type;
+  // typedef Lazy_exact_nt<ET>  Type;
+};
+template < typename ET >
+struct Exact_ring_selector<Lazy_exact_nt<ET> >
+: Exact_ring_selector<ET>
+{};
+
+} } //namespace CGAL::internal
 
 namespace Eigen {
   template<class> struct NumTraits;
